fix(economy): repair lib-economy test targets after u64→u128 migration
All five test targets failed to compile: amounts, fees and balances moved
from u64 to u128 in production but the tests were never updated. Nothing
in the crate has been tested since.

Two real bugs were hiding behind the broken build:

- EconomicModel::max_supply became u128::MAX, but get_economic_stats()
  serialises it via serde_json::json!, whose numbers top out at u64.
  Every call panicked with 'number out of range'. u128 amounts are now
  emitted as decimal strings.
- WalletBalance::can_afford is a plain >= check, so the assertion that a
  max-balance wallet cannot afford 1 was simply wrong.

Test amounts are capped at u64::MAX as u128 rather than u128::MAX to
preserve their pre-migration values; raising them trips an unchecked
multiply in calculate_fee (tracked separately).

180 tests now pass across all five targets.

Closes #2871

+++ lib-economy/src/wallets/transaction_history.rs
@@ -983,7 +983,7 @@ mod tests {
             let tx = create_payment_transaction(
                 [i; 32],
                 [i + 1; 32],
-                1000 * (i as u64 + 1),
+                1000 * (i as u128 + 1),
                 Priority::Normal,
             )
             .unwrap();
